Use ELSE [NULL] in prep for non-null branch fails
There's a new idea of making control or loop constructs that do not
run their branch/body return a ~void~ isotope.

This is contentious with the pattern:

    x: if condition [<value>]

Because if condition is false it will leave x in an unset state.  This
can be mitigated with OPT:

    x: opt if condition [<value>]

However, it is now the case with ~null~ isotopes that we are able to
write:

    x: if condition [<value>] else [null]

It actually looks a little more coherent that way.