Eliminate RETURN() macro, add return_throws
This makes use of the previous commit's ability to return cells not
in the frame output to make returns cleaner, avoiding the need for
the RETURN() macro.

It also adds a hook for `return_thrown` and renames the previous
`RETURN_INVISIBLE` to lowercase.

It gives them non-capitalized names to make them look more like
keywords, though there's no reasonable way to avoid using the
parentheses around the argument.  Hacks were considered, e.g.

    #define return_invisible \
        return Return_Invisible_Helper(D_OUT) ? nullptr :

This doesn't allow the routine to check anything about the passed
in value, however.  And is kind of just asking for trouble.