Allow return cells not in frame output
As a performance point, the system tried to discourage the return
of cells not in D_OUT (frame_->out) when running a native, as that
made the action dispatcher have to detect the case and copy it.

This gave rise to a RETURN() macro, later incarnated as RETURN_COPY()

Since the first check of the return result is for f->out, this
means those who are concerned about the micro-optimization can just
build values directly in f->out.  But for readability it just looks
better to be able to `return any_cell;`

Also fixes a bug caught by turning the parameter taken to be a
const REBVAL*.