Make reuse of dispatch of Generics easier
The Generic_Dispatcher allows the generic handler to return API
values or cells not in the frame output.  Reusing that means
each caller has to deal with those cases, and QUOTED! wasn't
getting it correct.

This makes a friendlier version that narrows the cases the reuser
of generic dispatch has to handle to just throwing or having the
result in the frame out.