Guard against recursive Panic()
Very large amounts of output gets dumped to the screen in a
recursive panic, and the original message can be scrolled off
even if you have a reasonably long scroll buffer in the terminal.

This exits on a recursive panic, but prints the file and line
number of the second panic.