Implement "Translucency" for void states
This is a major change of several inter-related things, that are
geared toward making things more coherent surrounding the handling
of "voids".

It introduces a new set of behaviors for "void isotopes", which are
non-reified notions of invisibility.  A void isotope only exists if
a chain of evaluations where it could have been invisible would have
yielded no intermediate non-void results:

    >> (~void~ comment "no values" ~void~)
    == ~void~  ; isotope

    >> (~void~ <some value here> ~void~)
    == ~  ; isotope (a.k.a. "none")

This reveals the non-reified nature of a void--which is in fact
only a flag sitting on top of whatever value or non-value was in
the evaluation chain before it.  It is only reified into a signal
of voids by ^META contexts which see a ~void~ BAD-WORD! and
thus report having seen a "~void~ isotope", even though no such
thing exists.

Because of this new mechanic, the ALL function can discard ~void~
isotopes without needing to store the previous result off to the
side.  Combined with making conditional constructs return ~void~
when no branch is taken, more useful behaviors arise:

    >> all [if true [1] if false [2]]
    == 1

Ultimately this culminates in a new useful signal of a branch
never been taken or of a loop body never running, which allows
the implementation of FOR-BOTH on top of two FOR-EACH loops, as
a demonstration of what can coherently be achieved with this.

Note: tests in this commit are still failing, but not crashing
and most represent semantic edges that will be best hashed out
by looking at real world experience.  So those tests will be
fixed in upcoming commits.