Allow GROUP! as variables container in loops
Historical Rebol did not allow plain groups to be evaluated as a
dereference for the variables to name for a loop.  You had to use
a GET-WORD!:

    varname: 'x
    foreach :varname [1 2 3] [print [x]]

Following this example, Ren-C made it so you had to use a GET-GROUP!
if you were going to have an array of code:

    for-each :(varname) [1 2 3] [print [x]]

In the attempt to make things look as palatable as possible in the
FOR-BOTH example, this goes ahead and hacks in support for plain
GROUP! in that slot.