Make dispatchers use OUT and SPARE
In order to get the optimized builds to not complain about use
of the `return_thrown` macro not using `frame_`, this goes ahead
and has the dispatchers use OUT and SPARE.  This keeps them from
using f->out and FRM_SPARE(f)...the latter of which led many to
create a local variable called "spare".

While doing that, this also makes return_thrown() accept arguments
other than OUT, and move them as-needed.  This cleans up callsites,
and also made it more obvious to see several places which were
actually returning the wrong cell.