Add FOR-BOTH test
FOR-BOTH was an early talking point for an extremely simple usermode
loop construct that would be built out of two FOR-EACH loops:

     >> for-both x [1 2] [3 4] [print [x], x * 10]
     1
     2
     3
     4
     == 40

Despite seeming like a simple thing to write in Redbol languages,
it's actually quite difficult to get a version that handles BREAK
and CONTINUE correctly, along with having the overall return result
be that of the last body evaluation.

This shows a working version that is elegant and builds on top of
loop invariants, and makes full use of the ^META facilities as well
as voids/invisibility:

    for-both: func ['var blk1 blk2 body] [
        unmeta all [
            meta for-each (var) blk1 body
            meta for-each (var) blk2 body
        ]
    ]

It includes tests that demonstrate its properties.