Align COMPOSE with errors for NULL, fix deep quoting
The new idea is that NULL--as representative of soft failure--
should not be quietly spliced via COMPOSE/REDUCE/etc...but
need to be triaged with MAYBE or TRY or REIFY.  This ties into
void taking over the role of "vanishing thing", as the indicator
of loops that did not run--or branching constructs where the
branch was not taken, etc.

This brings COMPOSE up to date with that, and fixes some issues
related to deep quoting discovered along the way.