Eliminate "pure" invisibility due to void isotopes
Pure invisibility originated from the idea of not wanting to make
every function using the evaluator worry about whether an eval step
disappeared or not, and have to erase it.  But realistically
speaking, this just pushed the difficulty around to where things
like UPARSE would have to know when things vanished anyway.

With the help of the ^META representation concept, ~void~ has
evolved into something that can exist in a reified meta form while
vanishing in plain form.  It is easier to handle by the constructs
that want to handle it, and streamlines things.

In the spirit of this change, `<invisible>` is changed to simply
`<void>` in function specs.