Change strategy for to-c-name, drop symbol hacks
This changes TO-C-NAME to not make symbols that start with underscore
then a capital letter (reserved in C)

It also works around the inability to create symbols with certain
names in the bootstrap executable by making the words list allow
the usage of...strings.  (Why this did not occur to me sooner is
unknown).

With that, the creation of special symbols at boot time isn't
necessary, as they exist already and can be accessed with Canon(...)