Eliminate SYMBOL! datatype, escape WORD!s w/|
There was an idea tried which was that things like . and / would
be "empty" PATH! and TUPLE! instances, corresponding to two blanks.
This extended to other cases like .. and // and ...

It was done to try and assist in the mechanics of doing things
like merging and joining paths, so that common code could handle
cases like //a and / by looking at the structure with blanks on
the ends.  Additionally, creating SYMBOL! as a second-class citizen
which couldn't be put in PATH! or TUPLE! helped avoid ambiguity.

This idea caused a fair bit of pain, so the new idea is to have
a mechanism for escaping weird words:

     >> path: 'obj.|weird/words.legal|

     >> type of path
     == #[datatype! path!]

     >> length of path
     == 2

     >> second path
     == |weird/words.legal|

     >> type of second path
     == #[datatype! word!]

The idea is that some words are marked as needing escaping all the
time, some when they have sigils attached, and some when they are
paths.  The case of / is that it only needs escaping when put in
a path.  The flags are set on the symbol entry itself, and then
leveraged when molding.