Use macros to shorten f->xxx access in evaluator
Observing the benefits on readability of using OUT and SPARE
instead of D_OUT and D_SPARE, this makes the evaluator use
the same macros OUT and SPARE in place of f->out and f_spare.

In the action executor it also redefines ARG and PARAM to
talk about the current argument and parameter, vs access by
name (which will never be relevant in generic evaluator code).