Reclaim unused EVAL_FLAG_XXX
Several evaluator flags were related to path dispatch, but that has
been completely unified with normal evaluation.  So there is no
such thing as a "PVS" frame.  All the flags used by that mode are
now obsolete, and can be reclaimed.

It also turned out to be unnecessary for the evaluator to know if the
requests above it were ^META or not.  Generally cleaner for it not
to know, anyway!

There is a flag for frames that blame their parents for errors, which
may have been a good usability idea.  Error reporting is rather not
friendly, so leaving that unused flag as a reminder in case it turns
out to be something that would help.