Per-frame Reb_State storage
One historical idea from Rebol is to make use of a number of "hot"
global buffers.  This is useful when the produced data is only
temporary, or when the precise size of an output is not known
in advance.  (For instance: a REDUCE can't accurately predict the
number of elements for the result block based on the number in the
input block, so pushing to a cached memory location first to get
an accurate count can avoid wasteful reallocations or unused memory
in the result.)

Stackless puts pressure on this, because these global buffers have
to be put in suspended animation in order to swap tasks.  This can
happen on any frame (in theory).

This does a relatively inefficient but simple change so that each
frame push will snapshot the start state of these buffers.  This
gives enough information that if the frame has to be suspended
it can know how much of that information to put aside.

More optimal approaches could be implemented--but this is geared
toward getting correctness first.