Shift handling of null/void branches to return
The "Do_Branch" code was taking responsibility for making sure
that branches did not return pure null or void isotopes.  This
drew a distinction between evaluating normally and evaluating
branches.

With stackless code handing off control of evaluation, it does
not make sense to have a different codebase for those evaluations
that is aware of the need to distort results from one that is.
Most likely it would be controlled by a flag, that corresponds to
a different kind of return intent.

This moves the handling into the return_branched() intent to pave
the way for that stackless model.