Kill off EVAL_FLAG_FULLY_SPECIALIZED
At one time, there was ambiguity whether NULL in a frame slot
meant the slot was literally NULL or if it was unspecialized.

Having isotopes--and them not being legal when running frames,
lets us choose the none isotope (~) as a reserved value to
mean unspecialized.  Any fields that are NULL can be taken
at face value as being intentional specializations with null.