Distinguish VOID from END, rethink bit patterns
The END Reb_Kind (REB_0) was used for two completely different
purposes.  One was as rebEND, the termination of a variadic
sequence in the API.  (At one time this also terminated arrays,
which was redundant with the length.  Now it no longer has a
termination signal, but enumerations are done to tail.)

The other use was for freshly initialized cells.  Detecting
that they had no content was how invisible evaluations were
noticed.

This switches it around so that the invalid UTF-8 byte 192
is used for end signals, and that the system does not set
cells to end signal.  This is not even marked as a cell--it
looks like a freed series.

Then, REB_0 is used in valid cell patterns that are both regular
and stale, to represent VOID.  This makes more sense in how
these cells are used than calling this state "END", and the
resulting code is clearer...with reduced risk of conflating
an end signal with an empty evaluation.

This also does a bit of work on the idea of leveraging the
ability to memset() a memory region to 0 and have it be valid
places to write cells to...and avoiding redundant assignments
to initialize already-initialized cells.