Cut down on _Maybe_Stale entry points
Prior to the unification of the stale flag and a cell's "free"
flag, staleness was a property that callers had to be very
careful with by contract--because ignoring the stale flag would
mean that something left over from a previous evaluation step
could be considered the output of a following step.  The need
to be careful led to the creation of routines with "_Maybe_Stale"
in their names as a reminder to check the result.

But once the stale flag tripped up read access, it was no
longer as risky.  Also, the limited number of routines that
actually want to see staleness could have an evaluation flag
saying so...with an assert that the incoming cell would be
void if it wasn't interested in staleness.

This switches to using the evaluation flag, and cuts down the
number of entry points significantly with the _Maybe_Stale()
variants coalesced with the regular versions.