Split up rebRunThrows and rebRunCoreThrows
There was a flag passed into rebRunThrows called "fully", which
represents only one of the flags you might want to pass to a
core-capable variadic routine.

This eliminates the "rebRunMaybeStaleThrows" in favor of a
version that takes arbitrary evaluation flags, and makes the
rebRunThrows() routine go ahead and default to requiring the
evaluation step make it to the end of the input--as that was
the most common need.

(rebRunCoreThrows() really shouldn't be part of the public API,
but being in %a-lib.c gets its variadic interface processed.)