Move branch result reification to EVAL_FLAG_BRANCH
In the stackless model, an evaluation frame can be handed over
with delegation, or set up as a continuation that will call the
function back.

It's cheaper if something like IF could delegate the evaluation
of its branch and not ask for a callback--just for the purpose
of setting NULL to a ~null~ isotope or voids to none.  This
motivates the creation of an EVAL_FLAG to capture that behavior,
so that the frame can be fully delegated by things like IF with
no need to be called back.