Allow ~end~ and ~void~ isotopes
While writing the SWITCH2 proposal, it was noticed that there
needs to be some way of having a FRAME! be formed from a feed
of values that is incomplete...and be able to make decisions
on what to do about that incompleteness.  Hence there would
need to be some kind of end marker.

This led to the realization that there is a place for reified
~end~ and ~void~ isotopes...in unfulfilled frame slots.

    >> make frame! make varargs! [add 1]
    == make frame! [
        return: '
        value1: 1
        value2: ~end~
    ]

Since isotopes are not truly legitimate values for frames that
are to be executed, they can serve as an instruction to the
type check phase later.  And this gives frame builders a chance
to change the state before executing, perhaps filling in the
missing values.

Once the isotopes exist, they have to have a meta state...so
it makes sense to give them the BAD-WORD! ones and move the true
void and end states to something different.  So they became
@end and @void.