Allow UPARSE ANY combinator to take plain BLOCK!
There was a bit of discomfort surrounding allowing the ANY
combinator to take a plain BLOCK! for a list of alternates,
since the block inside didn't follow the same rules as the
BLOCK! combinator in the rest of parse.

It's kind of silly to get hung up about that, considering it
is pretty much the premise of Rebol that everything gets
interpreted contextually.  Plain blocks look much better and
are learnable.