New naming trick: noquote(const Cell*)
A cell can actually report that it is a REB_QUOTED when its true
type is something else.  This is efficient, but it can lead to
confusion when this dichotomy of a single address having two
datatypes (the underlying type and the potentially quoted type)
wind up being conflated.

A trick for checking cases where this might be getting messed up
had a fairly lousy-looking name that did not communicate much.
It was a REBCEL differentiated from a RELVAL, when the name
"Cell" is being eyed for the concept of an array cell that has
not been Derelativized into a full Value with a known binding.

This flips the trick from REBCEL(const*) to noquote(const Cell*).
The result is more communicative, and allows the protective
type to share the name "Cell" with the relative value form.
Readers can know this is a nonquoted view of a cell.

The actual truth is that some places in the code use noquote
incorrectly at this time, which was pointed out when an assert
enforcing this was added to try and protect against REB_QUOTED
cell types.  But a simplification is forthcoming that will
get rid of those cells.