Eliminate KIND3Q_BYTE Complexity
Prior to the existence of series "flavors" and cell node bits,
there was a compression strategy which put higher-escaped
levels of quoting into separate cells.  This was chosen instead
of using a byte in the header to store the quote level because
there was a wish that PATH!s that acted like refinements would
be able to fit in a cell...so a byte was used to cover the
"inner type" of cell format that an aliased form was using.
Some type system trickery was used to keep the resulting
conflation from making things crash and burn.

However, the detection of a cell's nature can be done through
looking at the nodes it points to.  For instance, a PATH! can
point to a symbol series node, or to an array node, or a pair
node, or to no node at all and just compress bytes.  This is
enough information to get the discernment that was achieved
before, and generally more direct and simple to understand.

This opens up being able to put the quote level as a byte in
the cell header.  This winds up making things radically simpler
because there were some pretty gnarly epicycles of problems
related to figuring out how binding should work when the payload
of a cell had been broken out into a separate series node due
to a higher quoting level.

However, the mechanics for discerning views of a cell that is
its unquoted sense vs. quoted sense are still valuable.