Consolidate Do_Branch() to always take condition
In stackless, each evaluation intent has to be expressed as
a returned frame that is pushed to do some evaluation and then
(potentially) call back.  There are helpful `continue_xxx` macros
that put together common intents.

But there are a lot of variations in intent:

    continue_catchable_xxx
    continue_uncatchable_xxx
    delegate_xxx

Having separate branch and branch_with entry points makes a bit
of a mess.  Since you can do a branch without an argument just
by passing an END, it seems better to do that for the no-argument
form vs having `branch_with` and `branch` forms.

This proactively goes ahead and does that with the pre-stackless
codebase.  To make it more palatable, END_CELL is renamed to
just END.