Don't use ^META conventions for thrown values
There was a situation where END cells were being held in the
TG_Thrown_Arg slot to indicate no throw was in progress.  When
END cells were retired from use, this switched to being void.
This created a problem whereby things like RETURN could not
indicate they wanted to throw a void.

At the time, the most sensible solution seemed to be to make
the CATCH/THROW mechanics speak in ^META terms and pass the
values that way.  But a better answer is to make the unused
state a stale void, and then the receiving code that catches
throws can be written in a more direct way.