Bring back copying return macro, add checks
It seemed like a good idea to make the native `return` accept
arbitrary value pointers, and then just copy from them.

BUT this makes it too easy to slip up and return a pointer to a
local variable.  Having a return_xxx macro that does the copy
prior to returning solves that.

Using a revert of the commit that took away RETURN(), this brings
it back as `return_value`, and bolsters some asserts in the
other return macros.