Use Abort_Frame() in throwing scenarios
There was inconsistent application of Abort_Frame() vs. Drop_Frame()
when throws were in play.  Aborting does roll back of state to
where it is balanced to its baseline, while Drop_Frame() asserts
that everything was restored to baseline by the native.

It's not clear if it's better to have two different entry points
vs. just have Drop_Frame() trigger off of whether or not the
evaluator is in a throwing state.  But so long as two different
routines exist, they should be used consistently.
Use Abort_Frame() in throwing scenarios

There was inconsistent application of Abort_Frame() vs. Drop_Frame()
when throws were in play.  Aborting does roll back of state to
where it is balanced to its baseline, while Drop_Frame() asserts
that everything was restored to baseline by the native.

It's not clear if it's better to have two different entry points
vs. just have Drop_Frame() trigger off of whether or not the
evaluator is in a throwing state.  But so long as two different
routines exist, they should be used consistently.