Stackless "Trampoline", and stackless Trap
This is a bare minimum set of changes to route all evaluations
through a "trampoline" which hooks each evaluation step.  This
can be used to implement continuations--but for the moment, it
focuses on being able to correctly handle cases where each
invocation of the evaluator introduces a new stackful level.
It's a good stress test for error handling and throws that need
to be able to make it past frames that are reified on the C
call stack.

Notably, this changes the setjmp/longjmp logic such that there
is one setjmp/longjmp trap per trampoline invocation.  Here in
this early version, that introduces a significant overhead...
but as routines are converted to stackless, it will be cheaper
as well as offer failure interception cheaply at every level
of the stack.