Remove clearing of NO_LOOKAHEAD with GROUP!
There was a clearing of a NO_LOOKAHEAD flag in group evals
that claimed it was to avoid an assert.  This likely predated
the current simpler non-retriggering GROUP! logic, which was
ultimately deemed in service of a broken goal of commenting
out arbitrary groups of code.

It (fairly obviously) causes this problem:

>> 1 * 2 + (3) * 4
== 14

>> 1 * 2 + 3 * 4
== 20