Typecheck ^META arguments in specialization
It's not clear whether specialization should do typechecking
itself or defer to the function execution.  It's faster to do it
in advance, but it may turn out to be some kind of layering
violation to do it.

But as long as it is typechecking, it should allow for ^META
parameters to be checked.  This runs through the same code as
used in the typechecking phase of the evaluator.