Encode isotope status with QUOTE_BYTE 255
Originally, the isotope status was a CELL_FLAG on BAD-WORD!.
But since being an isotope is mutually exclusive with being
quoted, it makes sense to take a special quoting byte value and
make that the signal.

Not only does this reclaim a generic bit in the cell, it also
helps make it more obvious when a mistake is made and an isotope
is being processed by code that isn't aware it has an isotope.