Use new rebEntrap() vs. console rebRescue()
At one time, the ENTRAP function would put results in a block
if they weren't errors, otherwise return an ERROR!.  This let
a single result offer an error or a meta value.

The block was replaced by quoting, and now the ENTRAP idea is
basically the concept in general of ^META.  A non-quoted error
is indicative of "error isotopes" in the new thinking.

Yet it was desired that there not be anything visible on the
stack above user code, so running a GROUP! of code was sandboxed
by a rebRescue() operation, that passes an arbitrary C function
in to be bracketed by the fail-catching mechanics.

In the stackless model, this application of rebRescue() does
not make much sense...because it's a single evaluation call
through the Trampoline, which has its own protection and
bracketing.  This makes it seem like a better idea to just
have an API entry point that does what ENTRAP does, which is
cleaner.

(This is being considered as distinct from what rebMeta() does,
because that will only be geared to trap one immediate error...
vs. a deep one bubbling up.  Deep errors will probably call a
panic() handler if not caught.)