Stylize FOR-EACH, MAP, EVERY as Stackless
Because the enumeration state code is not yet stackless, the
iterating constructs can't be.  But the Loop_Each() driver that
was previously having a "loop mode" passed into it can be
stylized in such a way that it calls back out into the frame
phase that called it.  It just returns when it is finished.

(A better story for stackless iteration and lock management is
needed, but this is an incremental step toward that.)