Cache `v` and `gotten` in frame for stackless
In the evaluator, there were 2 temporary local variables which
were set during dispatch to be the currently switch()'d upon
value and its (potentially) fetched value if it was a word.

These would be held independently from the feed, because some
cases (like SET-WORD!) would arbitrarily advance the feed past
where it was at the moment the set-word was seen...but the
assignment had to happen later.

In stackless, it's not an option to hold these as local variables
in the "Evaluator_Executor()", because it gets kicked off the
stack.  So they have to be inside the heap-allocated frame
structure itself.