Opt into group evals in SET+GET w/o var
It could be surprising if a simple SET or GET operation ran
arbitrary code, due to passing in a GROUP! (or a TUPLE! or PATH!
with groups in it).  So when the idea that you should be able
to ask a SET or GET operation for a persistent set of "steps"
to identify came along, it was unified with the idea that if
you didn't ask then group evaluations would raise an error.

You could therefore "opt in" without naming a variable by
providing a blackhole (#).  But this provision worked only
at the SET and GET natives level.  The internal routines still
expected you to provide a variable--so scratch space (e.g. SPARE)
was used.

But since sometimes scratch space is hard to come by (especially
in stackless) it helps to have another indicator.  This uses the
END cell's address for GROUPS_OK to tell the internal routines
that it's all right to evaluate groups, but then there's no
variable backing it.  So it can make the operations more
efficient without complicating the interface with an orthogonal
flag.