Correctly use INERT_OPTIMIZATION flag in multi-return
The multi-return code was trying to suppress an assert which
pops up when a lookback is "too late".  It reacted to that
assert by setting the EVAL_FLAG_INERT_OPTIMIZATION...but it
did not use the semantics of the inert optimization (e.g. it
called the enfix lookahead regardless of whether it was
needed due to a WORD! in the next slot or not).

This adjusts that, and properly clears off the inert optimization
flag between evaluations--vs. having it get set on an evaluator
frame and stay stuck there (allowing too-late lookbacks that
come after the flag is set once to pass unnoticed).