Update evaluator exit flag checking
The checking of flags on evaluator exit has been a helpful way
to notice when things that shouldn't remain set are still set.
This helps reuse frames (e.g. EVAL_FLAG_TO_END being reused
across expressions) and also unexpected flags being set can
find missed optimizations or misunderstandings.

But the previously-used method would capture the "initial flags"
into a stack variable, and then make sure they hadn't changed.
A stack variable can't be used--so it would have to be a frame
variable.

How useful checking the flags haven't changed is questionable,
as the main thing is just about what flags are left set in an
absolute sense.  This changes the check, and also has it print
out the bit numbers of stray flags set (very helpful for
understanding the meaning of the results).