Fix QUIT-WHEN-DONE feature of MAIN-STARTUP
The quit-when-done was emitting a QUIT 0 command and then
giving `<unreachable>` as a TAG! signal to RETURN.  But that
signal was not understood.  So the code was working for some
random reason.

This makes the signal `<quit>` instead, which makes things work
with some changes to return values.  The design needs review
(and will get one in stackless...)