Enhance LAMBDA to allow for spec blocks
This changes the LAMBDA definition so that you can give functions
full spec blocks, refinements, etc.  It still prohibits a RETURN:
spec, because in the current scheme of thing that implies a
RETURN: function.

(Note: This might be loosened, to say that a LAMBDA with a RETURN
has an implicit call to RETURN in it so the body is checked.)