FUNC and FUNCTION enforce RETURN usage
For reasons both aesthetic and cognitive, it's believed that a
function which returns a result should actively use RETURN to
mark that it is returning values--vs. have them fall out of the
bottom of the function implicitly.

Rather than erroring, a plain FUNC(TION) will return a none isotope
if no RETURN is at the end of the body.  This gives parity with
the behavior of JavaScript functions.

For a function-like construct that implicitly evaluates to its
body result, there is the option of using LAMBDA--which has no
RETURN construct of its own.