Make port actors actually use LAMBDA or FUNCTION
For reasons that are not clear, R3-Alpha did some kind of
analysis of the port actor block to make sure it had FUNC as
the function generator.  Modifying this code to tolerate LAMBDA
needed to also make it actually use that generator instead of
just allowing LAMBDA and then calling it a FUNC anyway.

(Note: none of this made any sense before, and it doesn't make
any sense now.  Putting PORT!/scheme code in the ReplPad was
done against my preference.)