Make prefix LAMBDA not quote its arguments
The infix form of LAMBDA needed to quote its left hand side to
get WORD!s.  But the prefix form of lambda is more likely to be
used with code that constructs a spec (or a spec in a variable).

This removes the quoting from the prefix form, and fixes the
invocation of generators in the port actor.