Reserve arity-0 RETURN for returning void
This takes the code that allowed a spec of `return: <none>` to
specialize REUTRN to give a none isotope when no argument was
given, and instead errors on no argument.

While people are free to redefine RETURN as they wish, the core
wants to set a good example of being clear that a RETURN with
no arguments generally is meant to imply void returns.