Permit LAMBDA (including -> form) to be invisible
Originally lambdas were not allowed to be invisible--when being
invisible was more likely to throw a wrench into the structure
of evaluated code.

But now that invisibles only act invisibly at "interstitial"
moments in code, it's less of a danger.  It means that there can
be benefits to ease of use writing a wrapper for an opportunistic
invisible function as a lambda.

   >> vanisher: x -> []

   >> 1 + 2 vanisher 20
   == 3

(There is also a performance advantage--in particular in stackless,
to not needing a step to come back and move the value from a
temporary cell to the output cell.)