Fix ADAPT Layering Violation Issue (circa 2018)
This resolves a lingering issue regarding the presence of the
definitional RETURN from FUNC/FUNCTION being seen by ADAPT.

The simple resolution is: *It's not there because the generator
providing it hasn't run yet.*

When people want the intent to return without running the function
being ADAPT'ed, what they need to do is use ENCLOSE.  ADAPT is for
massaging the input frame and then running the adapted function
precisely one time.  For other intents (like possibly 0, or 10,
or however many times) then ENCLOSE must be used.