Drop optimizations on FUNC variants
In the early creation of "dispatchers", it was tempting to make
optimized dispatchers to cover various things.  TYPECHECKER was
one that is very useful today--that makes things like INTEGER?
faster than they would be if they had to be built up out of
things like EQUAL? to a TYPE OF.

But in addition to those high-value dispatchers, there were some
micro-optimizations applied to functions that didn't have bodies
that saved on things like the execution of an empty block, e.g.

    foo: func [a] []  ; it would optimize this, to not `do []`

Arguments were made that these cases were not uncommon because
they could be stub behavior for something given a behavior via
HIJACK (e.g. ASSERT).  But even so, the complexity of maintaining
the variants wasn't worth the advantage given.

The introduction of LAMBDA gave a more meaningful optimization,
especially in stackless--of defining a function class that has
no RETURN and hence no need for evaluating into an alternate
cell that needs to be copied to.  (More importantly, not needing
a phase of being revisited after a body evaluation at all.)