Rethink THROW to not overwrite output cells
At the outset, a THROW was written to the output cell as part of
a protocol to say that whatever result had been there previously
was now no longer valid.  When labeled throws were added to
Ren-C, it was thought that putting the label into the output
cell instead of the thrown value could help be more "disruptive"
and make it more likely for callers to notice that it didn't hold
the value they expected.

New techniques involving CELL_FLAG_STALE let us strategically mark
a cell as not readable by normal routines.  Though this was made
initially to prevent enfix operators from using evaluative results
from operations that have been "stepped past", it can apply well
to thrown results too--stopping them from being used.

Another issue is that by overwriting the output, it became not
possible to THROW or UNWIND a void value...because the cell the
thrown result was being written into might hold the prior output.
The stale flag works well for this purpose too.

So this shifts the concept of throwing to be something done on
a frame basis, not a cell-basis.  (In actuality it uses globals,
but the FRAME is used as a placeholder to say that it isn't
something necessarily global--if there were multiple stacks
running on parallel threads--if there ever were.)