Remove UNWIND checking from Func_Dispatcher()
With the old model of THROW using the output cell to pass up the
label of the throw, it was not possible to do an UNWIND of an
invisible result...because the signal of the UNWIND itself would
overwrite the stale result that was desired as output.

This meant that the Func_Dispatcher() wrote into a stale cell,
but had to catch any UNWIND throws before passing them up so
that the output cell wasn't overwritten.  This replicated work
done in the action dispatcher for catching unwind requests to
other functions (such as natives).

That is no longer needed.  It's still necessary for the Func's
dispatcher to write to a stale cell (since it wants to be able
to return invisibly even when there's evaluative products in the
body.)  But it can pass up a void UNWIND signal to the generic
unwind handling without worrying about that overwriting the
stale output.