Replace `return_void (OUT)` with `return VOID`
The ability to return void is limited very narrowly to frame
output cells (as there is no "void value" in the system that
represents the notion of voidness--that's paradoxical).

It's desirable to check that conditions to allow voidness are
met before returning from a native, so it was done with a macro
called `return_void`.  It took an output cell--but this cell
had to be the frame's OUT, as no other cell can be an indicator
of void.

Given these constraints, a better idea was thought of to just
say that the macros for testing and returning the R_VOID result
would be put together in a macro called VOID, so it could say
simply `return VOID;`

This is contentious with Windows defining VOID in the global
space as an oh-so-useful synonym for C's `void`.  :-/  But it's
already decided not to let Windows dictate the terms of the
evaluator, so this goes along with OUT as a define taken and
then the ability to not use the shorthands when including
%sys-core.h if absolutely necessary.