Stackless FOR-EACH, MAP-EACH, and EVERY
This takes the existing ugly code and makes it into an iterator.

(There are a lot of outstanding issues pertaining to locking, and
this is really just carrying forward a bunch of chaos from R3-Alpha.
But it's hopefully an interim step toward something better...and can
be at least good enough to not need Asyncify in the web build.)