Change Is_Action_Frame() to test for Action_Executor
Before the existence of generalized Executor functoins for frames,
the way to tell if something was an action frame was if there
was an action pointer in the f->original slot.  Any non-action
frame would guarantee that was nullptr, and any action frame
would guarante it was non-nullptr.

With the executors, it's obviously better to test if it's using
the Action_Executor.  This means f->original can be uninitialized
when a frame is created.